% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{comment}
\alias{comment}
\title{Add a generic comment to the dtrackr history graph}
\usage{
comment(
  .data,
  .messages = .defaultMessage(),
  .headline = .defaultHeadline(),
  .type = "info",
  .asOffshoot = (.type == "exclusion")
)
}
\arguments{
\item{.data}{- a dataframe which may be grouped}

\item{.messages}{- a character vector of glue specifications. A glue specification can refer to any grouping variables of .data, or any variables defined in the calling environment, the \{.total\} of all rows,
the \{.count\} variable which is the count in each group and \{.strata\} a description of the group}

\item{.headline}{- a glue specification which can refer to grouping variables of .data, or any variables defined in the calling environment, or the \{.total\} variable which is nrow(.data)and \{.strata\}}

\item{.type}{- one of "info","...,"exclusion": used to define formatting}

\item{.asOffshoot}{- do you want this comment to be an offshoot of the main flow (default = FALSE).}
}
\value{
the same .data dataframe with the history graph updated
}
\description{
Add a generic comment to the dtrackr history graph
}
\examples{
iris \%>\% p_comment("hello {.total} rows")
}
