% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{distinct.trackr_df}
\alias{distinct.trackr_df}
\title{Distinct values of data}
\usage{
\method{distinct}{trackr_df}(
  .data,
  .f,
  ...,
  .keep = FALSE,
  .messages = "removing {.count.in-.count.out} duplicates",
  .headline = "{.strata}"
)
}
\arguments{
\item{.data}{- a dataframe which may be grouped}

\item{.f}{a function as expected by dplyr::group_modify e.g. function(d,g,...) { ...do something with d and return a dataframe... }}

\item{...}{additional parameters for .f.}

\item{.keep}{- are the grouping variables kept in d, or split out to g (the default)}

\item{.messages}{- a set of glue specs. The glue code can use any global variable, or {.strata},{.count.in},and {.count.out}}

\item{.headline}{- a headline glue spec. The glue code can use any global variable, or {.strata},{.count.in},and {.count.out}}
}
\value{
the transformed .data.
}
\description{
Distinct acts in the same way as DPLYR. Prior to the operation the size of the group is calculated {.count.in} and
after the operation the output size {.count.out}
The group {.strata} is also available (if grouped) for reporting
}
\examples{
p_bind_rows(iris, iris \%>\% p_filter(Petal.Length > 5)) \%>\%
   p_group_by(Species) \%>\% p_distinct() \%>\% p_get()
}
