% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{dot2svg}
\alias{dot2svg}
\title{Convert dot to svg}
\usage{
dot2svg(dot)
}
\arguments{
\item{dot}{- a graphviz dot string}
}
\value{
the SVG as a string
}
\description{
Convert a digraph as string to SVG as string
}
\examples{
dot2svg("digraph {A->B}")
}
