% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr-package.R
\docType{package}
\name{dtrackr-package}
\alias{dtrackr}
\alias{dtrackr-package}
\title{dtrackr: Track your Data Pipelines}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Track and document 'dplyr' data pipelines. As you filter, mutate, and join your way through a data set, 'dtrackr' seamlessly keeps track of your data flow and makes publication ready documentation of a data pipeline simple.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://terminological.github.io/dtrackr/index.html}
  \item \url{https://github.com/terminological/dtrackr}
  \item Report bugs at \url{https://github.com/terminological/dtrackr/issues}
}

}
\author{
\strong{Maintainer}: Robert Challen \email{rob.challen@bristol.ac.uk} (\href{https://orcid.org/0000-0002-5504-7768}{ORCID})

}
\keyword{internal}
