% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{flowchart}
\alias{flowchart}
\title{Flowchart output}
\usage{
flowchart(
  .data,
  filename = NULL,
  size = std_size$half,
  maxWidth = size$width,
  maxHeight = size$height,
  rot = size$rot,
  formats = c("dot", "png", "pdf", "svg"),
  ...
)
}
\arguments{
\item{.data}{- the tracked dataframes}

\item{filename}{- a filename (without extension) which will be where the formatted flowcharts are saved}

\item{size}{- a list of length and width in inches e.g. a std_size}

\item{maxWidth}{- a width in inches is size is not defined}

\item{maxHeight}{- a heught in inches if size is not defined}

\item{rot}{- an angle of rotation for the saved file if size is not defined}

\item{formats}{- some of "pdf","dot","svg","png","ps"}

\item{...}{- other params passed onto p_get_as_dot, notable ones are fill, fontsize, colour, size, maxWidth and maxHeight}
}
\value{
the output depends on the context in which the function is called. Some form of browse-able output if in an interactive session or a PNG/PDG link if in knitr
}
\description{
Create a flowchart of the history of the dataframe
}
