% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ILPD}
\alias{ILPD}
\title{Indian Liver Patient Dataset}
\format{
A data frame with 583 rows and 11 variables:
\describe{
\item{Age}{}
\item{Gender}{}
\item{Total_Bilirubin}{}
\item{Direct_Bilirubin}{}
\item{Alkaline_Phosphatase}{}
\item{Alamine_Aminotransferase}{}
\item{Aspartate_Aminotransferase}{}
\item{Total_Protein}{}
\item{Albumin}{}
\item{Albumin_Globulin_Ratio}{}
\item{Case_or_Control}{Selector field used to split the data into two sets (labeled by the experts)}
}
}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/ILPD+(Indian+Liver+Patient+Dataset)}
}
\usage{
ILPD
}
\description{
This data set contains 416 liver patient records and 167 non liver patient records.
The data set was collected from north east of Andhra Pradesh, India. Selector is a
class label used to divide into groups(liver patient or not). This data set contains
441 male patient records and 142 female patient records.
}
\details{
\enumerate{
\item Bendi Venkata Ramana, Prof. M. S. Prasad Babu and Prof. N. B. Venkateswarlu, A Critical Comparative Study of Liver Patients from USA and INDIA: An Exploratory Analysis, International Journal of Computer Science Issues, ISSN :1694-0784, May 2012.
\item Bendi Venkata Ramana, Prof. M. S. Prasad Babu and Prof. N. B. Venkateswarlu, A Critical Study of Selected Classification Algorithms for Liver Disease Diagnosis, International Journal of Database Management Systems (IJDMS), Vol.3, No.2, ISSN : 0975-5705, PP 101-114, May 2011.
\item Dua, D. and Graff, C. (2019). UCI Machine Learning Repository \url{http://archive.ics.uci.edu/ml/}. Irvine, CA: University of California, School of Information and Computer Science.
}
}
\keyword{datasets}
