% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{landscape}
\alias{landscape}
\title{Convert page size from portrait to landscape}
\usage{
landscape(size)
}
\arguments{
\item{size}{\itemize{
\item list of width and height in inches, e.g. a std_size
}}
}
\value{
a landscape size
}
\description{
Convert page size from portrait to landscape
}
