% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_capture_exclusions}
\alias{p_capture_exclusions}
\title{Start capturing exclusions on a tracked dataframe.}
\usage{
p_capture_exclusions(.data, .capture = TRUE)
}
\arguments{
\item{.data}{- a tracked dataframe}

\item{.capture}{- Should we capture exclusions (things removed from the data set). This is useful for debugging data issues but comes at a significant cost. Defaults to the value of getOption("dtrackr.exclusions") or FALSE.}
}
\value{
the .data dataframe with the exclusions flag set or cleared.
}
\description{
Start capturing exclusions on a tracked dataframe.
}
\examples{
iris \%>\% track() \%>\% capture_exclusions()
}
