% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr.R
\name{p_comment}
\alias{p_comment}
\title{Add a generic comment to the trackr history graph}
\usage{
p_comment(
  .data,
  .messages = NULL,
  .headline = NULL,
  .type = "info",
  .asOffshoot = FALSE
)
}
\arguments{
\item{.data}{- a dataframe which may be grouped}

\item{.messages}{- a character vector of glue specifications. A glue specification can refer to any grouping variables of .data, or any variables defined in the calling environment, or the {.count} variable which is nrow(.data)}

\item{.headline}{- a glue specification which can refer to grouping variables of .data, or any variables defined in the calling environment, or the {.total} variable which is nrow(.data)}

\item{.type}{- one of "info","...,"exclusion": used to define formatting}

\item{.asOffshoot}{- do you want this comment to be an offshoot of the main flow (default = FALSE).}
}
\value{
the same .data dataframe with the history graph updated
}
\description{
Add a generic comment to the trackr history graph
}
\examples{
iris \%>\% p_comment("hello {.total} rows")
}
