% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_copy}
\alias{p_copy}
\title{Copy the dtrackr history graph from one df to another}
\usage{
p_copy(.data, from)
}
\arguments{
\item{.data}{- a dataframe which may be grouped}

\item{from}{- the dataframe to copy the history graph from}
}
\value{
the .data with the history graph of "from"
}
\description{
Copy the dtrackr history graph from one df to another
}
\examples{
mtcars \%>\% p_copy(iris \%>\% p_comment("A comment"))
}
