% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_count_if}
\alias{p_count_if}
\title{Simple count_if dplyr summary function}
\usage{
p_count_if(...)
}
\arguments{
\item{...}{- expression to be evaluated}
}
\value{
summary count
}
\description{
Simple count_if dplyr summary function
}
\examples{
iris \%>\% p_group_by(Species) \%>\% p_mutate(long_ones = p_count_if(Petal.Length > 4))
}
