% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_count_subgroup}
\alias{p_count_subgroup}
\title{Add a subgroup count to the dtrackr history graph}
\usage{
p_count_subgroup(
  .data,
  .subgroup,
  ...,
  .messages = .defaultCountSubgroup(),
  .headline = .defaultHeadline(),
  .type = "info",
  .asOffshoot = FALSE,
  .tag = NULL,
  .maxsubgroups = .defaultMaxSupportedGroupings()
)
}
\arguments{
\item{.data}{- a dataframe which may be grouped}

\item{.subgroup}{- a column with a small number of levels (e.g.)}

\item{...}{- additional parameters will be passed to factor(subgroup,...) to control levels, ordering, etc.}

\item{.messages}{- a character vector of glue specifications. A glue specification can refer to anything from the calling environment and .name for the subgroup name, .count for the subgroup count, .subtotal for the current grouping count and .total for the whole count}

\item{.headline}{- a glue specification which can refer to grouping variables of .data, or any variables defined in the calling environment}

\item{.type}{- one of "info","exclusion": used to define formatting}

\item{.asOffshoot}{- do you want this comment to be an offshoot of the main flow (default = FALSE).}

\item{.tag}{- if you want the summary data from this step in the future then give it a name with .tag.}

\item{.maxsubgroups}{- the maximum number of discrete values allowed is configurable with `options("dtrackr.max_supported_groupings"=XX)`. The default is 16.}
}
\value{
the same .data dataframe with the history graph updated
}
\description{
Because of the ... summary specification parameters MUST BE NAMED.
}
\examples{
ILPD \%>\% p_group_by(Case_or_Control) \%>\%
   p_count_subgroup(
      .subgroup = Gender,
      .messages="{.name}: {.count}/{.subtotal}",
      .headline="{Case_or_Control}: {.subtotal}/{.total}"
   ) \%>\% p_get()
}
