% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_excluded}
\alias{p_excluded}
\title{Get the dtrackr excluded data record}
\usage{
p_excluded(.data, simplify = TRUE)
}
\arguments{
\item{.data}{- a dataframe which may be grouped}

\item{simplify}{- return a single summary dataframe of all exclusions.}
}
\value{
a dataframe of the excluded data. This has a nested structure containing records excluded at each part of the pipeline.
}
\description{
Get the dtrackr excluded data record
}
\examples{
iris \%>\% p_excluded()
}
