% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_filter}
\alias{p_filter}
\title{Filtering data}
\usage{
p_filter(
  .data,
  ...,
  .preserve = FALSE,
  .messages = "excluded {.excluded} items",
  .headline = .defaultHeadline(),
  .type = "exclusion",
  .asOffshoot = (.type == "exclusion"),
  .stage = "",
  .tag = NULL
)
}
\arguments{
\item{.data}{a dataframe which may be grouped}

\item{...}{the filter criteria}

\item{.preserve}{Relevant when the \code{.data} input is grouped.
If \code{.preserve = FALSE} (the default), the grouping structure
is recalculated based on the resulting data, otherwise the grouping is kept as is.}

\item{.messages}{a set of glue specs. The glue code can use any global
variable, or \{.strata\},\{.count.in\},and \{.count.out\}}

\item{.headline}{a headline glue spec. The glue code can use any global
variable, or \{.strata\},\{.count.in\},and \{.count.out\}}

\item{.type}{the format type of the action typically an exclusion}

\item{.asOffshoot}{if the type is exclusion, \code{.asOffshoot} places the
information box outside of the main flow, as an exclusion.}

\item{.stage}{a name for this step in the pathway}

\item{.tag}{if you want the summary data from this step in the future then
give it a name with \code{.tag}.}
}
\value{
the filtered .data dataframe with history graph updated
}
\description{
Filter acts in the same way as in \code{dplyr} where predicates which evaluate to
TRUE act to select items to include, and items for which the predicate cannot
be evaluated are excluded. For tracking prior to the filter operation the
size of each group is calculated \{.count.in\} and after the operation the
output size of each group \{.count.out\}. The grouping \{.strata\} is also
available (if grouped) for reporting. See \code{\link[dplyr:filter]{dplyr::filter()}}.
}
\examples{
library(dplyr)
library(dtrackr)
tmp = iris \%>\% track() \%>\% group_by(Species)
tmp \%>\% filter(Petal.Length > 5) \%>\% history()
}
\seealso{
dplyr::filter()
}
