% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr.R
\name{p_get_as_dot}
\alias{p_get_as_dot}
\title{DOT output}
\usage{
p_get_as_dot(.data, fill = "lightgrey", fontsize = "8", colour = "black", ...)
}
\arguments{
\item{.data}{- the tracked dataframe}

\item{fill}{- the default node fill colour}

\item{fontsize}{- the default font size}

\item{colour}{- the default font colour}

\item{...}{- not used}
}
\value{
a dot string
}
\description{
Outputs a trackr history graph as a DOT string for rendering with graphviz
}
