% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_pause}
\alias{p_pause}
\title{Pause tracking the dataframe}
\usage{
p_pause(.data)
}
\arguments{
\item{.data}{- a tracked dataframe}
}
\value{
the dataframe paused
}
\description{
Pause tracking the dataframe
}
