% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_rename_with}
\alias{p_rename_with}
\title{Standard dplyr modifying operations}
\usage{
p_rename_with(.data, ..., .messages = "", .headline = "")
}
\arguments{
\item{.data}{- a dataframe which may be grouped}

\item{...}{a set of dplyr summary espressions.}

\item{.messages}{- a set of glue specs. The glue code can use any global variable, grouping variable, or {.strata}. Defaults to nothing.}

\item{.headline}{- a headline glue spec. The glue code can use any global variable, grouping variable, or {.strata}. Defaults to nothing.}
}
\value{
the .data but dplyr::ungrouped.
}
\description{
Equivalent Dplyr functions for mutating, selecting and renaming a data set act in the normal way.
mutates / selects / rename generally don't add anything in documentation so the default behaviour is to miss these out of the history.
This can be overridden with the .messages, or .headline values
}
\examples{
iris \%>\% p_group_by(Species) \%>\%
    p_mutate(Petal.Length = 2*Petal.Length, .messages="doubling length") \%>\%
    p_get()
}
