% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_resume}
\alias{p_resume}
\title{Resume tracking the dataframe. This may reset the grouping of the tracked data}
\usage{
p_resume(.data)
}
\arguments{
\item{.data}{- a tracked dataframe}
}
\value{
the dataframe resumed
}
\description{
Resume tracking the dataframe. This may reset the grouping of the tracked data
}
