% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_set}
\alias{p_set}
\title{Set the dtrackr history graph}
\usage{
p_set(.data, .graph)
}
\arguments{
\item{.data}{- a dataframe which may be grouped}

\item{.graph}{- a history graph list (consisting of nodes, edges, and head) see examples}
}
\value{
the .data dataframe
}
\description{
Set the dtrackr history graph
}
\examples{
mtcars \%>\% p_set(iris \%>\% p_comment("A comment") \%>\% p_get())
}
