% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_status}
\alias{p_status}
\title{Add a summary to the dtrackr history graph}
\usage{
p_status(
  .data,
  ...,
  .messages = .defaultMessage(),
  .headline = .defaultHeadline(),
  .type = "info",
  .asOffshoot = FALSE
)
}
\arguments{
\item{.data}{- a dataframe which may be grouped}

\item{...}{- any normal dplyr::summarise specification, e.g. count=n() or av = mean(x) etc.}

\item{.messages}{- a character vector of glue specifications. A glue specification can refer to the summary outputs, any grouping variables of .data, the .strata, or any variables defined in the calling environment}

\item{.headline}{- a glue specification which can refer to grouping variables of .data, or any variables defined in the calling environment}

\item{.type}{- one of "info","exclusion": used to define formatting}

\item{.asOffshoot}{- do you want this comment to be an offshoot of the main flow (default = FALSE).}
}
\value{
the same .data dataframe with the history graph updated
}
\description{
Because of the ... summary specification parameters MUST BE NAMED.
}
\examples{
iris \%>\% p_group_by(Species) \%>\%
   p_status(
      long = p_count_if(Petal.Length>5),
      short = p_count_if(Petal.Length<2),
      .messages="{Species}: {long} long ones & {short} short ones"
   ) \%>\% p_get()
}
