% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr.R
\name{p_summarise}
\alias{p_summarise}
\title{Summarise a data set}
\usage{
p_summarise(
  .data,
  ...,
  .groups = NULL,
  .messages = NULL,
  .headline = "{.strata}"
)
}
\arguments{
\item{.data}{- a dataframe which may be grouped}

\item{...}{a set of dplyr summary expressions.}

\item{.groups}{- Experimental lifecycle Grouping structure of the result.}

\item{.messages}{- a set of glue specs. The glue code can use any summary variable defined in the ... parameter, or any global variable, or {.strata}}

\item{.headline}{- a headline glue spec. The glue code can use any summary variable defined in the ... parameter, or any global variable, or {.strata}}
}
\value{
the .data but ungrouped.
}
\description{
Summarising a data set acts in the normal way. Any columns resulting form the summary can be added to the history graph
In the history this joins any stratified branches and acts as a specific type of p_summary, allowing you to
generate some summary statistics about the un-grouped data
}
\examples{
iris \%>\%
   p_group_by(Species) \%>\%
   p_summarise(avg = mean(Petal.Length), .messages="{avg} length") \%>\%
   p_get()
}
