% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr.R
\name{p_track}
\alias{p_track}
\title{Strart tracking the trackr history graph}
\usage{
p_track(.data)
}
\arguments{
\item{.data}{- a dataframe which may be grouped}
}
\value{
the hisotry graph
}
\description{
Strart tracking the trackr history graph
}
\examples{
iris \%>\% p_track()
}
