% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_ungroup}
\alias{p_ungroup}
\title{Remove a stratification from a data set}
\usage{
p_ungroup(x, ..., .messages = "total {.count} items", .headline = NULL)
}
\arguments{
\item{x}{- a dataframe which may be grouped (why not .data?)}

\item{...}{a set of dplyr summary expressions. If left blank a default of ".count=n()" will be filled in}

\item{.messages}{- a set of glue specs. The glue code can use any summary variable defined in the ... parameter, or any global variable, or {.count}. the default is "total {.count} items"}

\item{.headline}{- a headline glue spec. The glue code can use any summary variable defined in the ... parameter, or {.count}.}
}
\value{
the .data but dplyr::ungrouped.
}
\description{
Un-grouping a data set logically combines the different arms.
In the history this joins any stratified branches and acts as a specific type of p_summary, allowing you to
generate some summary statistics about the un-grouped data
}
\examples{
iris \%>\%
   p_group_by(Species) \%>\%
   p_comment("A test") \%>\%
   p_ungroup(
      avg = mean(Petal.Length),
      count = dplyr::n(),
      .messages="{count} items with {avg} petal length"
   ) \%>\%
   p_get()
}
