% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_untrack}
\alias{p_untrack}
\title{Remove tracking from the dataframe}
\usage{
p_untrack(.data)
}
\arguments{
\item{.data}{- a tracked dataframe}
}
\value{
the dataframe with no history
}
\description{
Remove tracking from the dataframe
}
