% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{pivot_longer.trackr_df}
\alias{pivot_longer.trackr_df}
\title{Reshaping data using \code{tidyr::pivot_longer}}
\usage{
\method{pivot_longer}{trackr_df}(data, ..., .messages = "", .headline = "", .tag = NULL)
}
\arguments{
\item{data}{A data frame to pivot.}

\item{...}{Additional arguments passed on to methods.}

\item{.messages}{a set of glue specs. The glue code can use any global
variable, grouping variable, or \{.strata\}. Defaults to nothing.}

\item{.headline}{a headline glue spec. The glue code can use any global
variable, grouping variable, or \{.strata\}. Defaults to nothing.}

\item{.tag}{if you want the summary data from this step in the future then
give it a name with .tag.}
}
\value{
the result of the \code{tidyr::pivot_longer} but with a history graph
updated.
}
\description{
A drop in replacement for \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}} which optionally takes a
message and headline to store in the history graph.
}
\seealso{
tidyr::pivot_longer()
}
