% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{print.trackr_graph}
\alias{print.trackr_graph}
\title{Print a history graph to the console}
\usage{
\method{print}{trackr_graph}(x, ...)
}
\arguments{
\item{x}{a dtrackr history graph (e.g. output from \code{\link[=p_get]{p_get()}})}

\item{...}{not used}
}
\value{
nothing
}
\description{
Print a history graph to the console
}
\examples{
library(dplyr)
iris \%>\% comment("hello {.total} rows") \%>\% history() \%>\% print()
}
