% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{ungroup}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr:group_by]{ungroup}}}
}}

