% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\docType{data}
\name{std_size}
\alias{std_size}
\title{Standard paper sizes}
\format{
An object of class \code{list} of length 8.
}
\usage{
std_size
}
\description{
A list of standard paper sizes
}
\keyword{datasets}
