% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{untrack}
\alias{untrack}
\title{Remove tracking from the dataframe}
\usage{
untrack(.data)
}
\arguments{
\item{.data}{a tracked dataframe}
}
\value{
the .data dataframe with history graph metadata removed.
}
\description{
Remove tracking from the dataframe
}
\examples{
library(dplyr)
library(dtrackr)
iris \%>\% track() \%>\% untrack() \%>\% class()
}
