% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{capture_exclusions}
\alias{capture_exclusions}
\title{Start capturing exclusions on a tracked dataframe.}
\usage{
capture_exclusions(.data, .capture = TRUE)
}
\arguments{
\item{.data}{\itemize{
\item a tracked dataframe
}}

\item{.capture}{\itemize{
\item Should we capture exclusions (things removed from the data set). This is useful for debugging data issues but comes at a significant cost. Defaults to the value of \code{getOption("dtrackr.exclusions")} or \code{FALSE}.
}}
}
\value{
the .data dataframe with the exclusions flag set (or cleared if \code{.capture=FALSE}).
}
\description{
Start capturing exclusions on a tracked dataframe.
}
\examples{
library(dplyr)
tmp = iris \%>\% track() \%>\% capture_exclusions()
tmp \%>\% filter(Species!="versicolor") \%>\% history()
}
