% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{dot2svg}
\alias{dot2svg}
\title{Convert \code{Graphviz} dot content to a SVG}
\usage{
dot2svg(dot)
}
\arguments{
\item{dot}{a \code{graphviz} dot string}
}
\value{
the SVG as a string
}
\description{
Convert a \code{graphviz} dot digraph as string to \code{SVG} as string
}
\examples{
dot2svg("digraph { A->B }")
}
