% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{filter.trackr_df}
\alias{filter.trackr_df}
\title{Filtering data}
\usage{
\method{filter}{trackr_df}(
  .data,
  ...,
  .preserve = FALSE,
  .messages = "excluded {.count.in-.count.out} items",
  .headline = "{.strata}",
  .type = "exclusion",
  .asOffshoot = (.type == "exclusion")
)
}
\arguments{
\item{.data}{- a dataframe which may be grouped}

\item{...}{the filter criteria}

\item{.preserve}{Relevant when the \code{.data} input is grouped.
If \code{.preserve = FALSE} (the default), the grouping structure
is recalculated based on the resulting data, otherwise the grouping is kept as is.}

\item{.messages}{- a set of glue specs. The glue code can use any global variable, or {.strata},{.count.in},and {.count.out}}

\item{.headline}{- a headline glue spec. The glue code can use any global variable, or {.strata},{.count.in},and {.count.out}}

\item{.type}{- the format type of the action - typically an exclusion}

\item{.asOffshoot}{- if the type is exclusion, asOffshoot places the information box outside of the main flow, as an exclusion.}
}
\value{
the transformed .data.
}
\description{
Filter acts in the same way as DPLYR. Prior to the operation the size of the group is calculated {.count.in} and
after the operation the output size {.count.out}
The group {.strata} is also available (if grouped) for reporting
}
\examples{
iris \%>\% p_group_by(Species) \%>\% p_filter(Petal.Length > 5) \%>\% p_get()
}
