% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_clear}
\alias{p_clear}
\title{Clear the dtrackr history graph}
\usage{
p_clear(.data)
}
\arguments{
\item{.data}{\itemize{
\item a dataframe which may be grouped
}}
}
\value{
the .data dataframe with the history graph removed
}
\description{
Clear the dtrackr history graph
}
\examples{
library(dplyr)
mtcars \%>\% track() \%>\% comment("A comment") \%>\% p_clear() \%>\% history()
}
