% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_clear}
\alias{p_clear}
\title{Clear the dtrackr history graph}
\usage{
p_clear(.data)
}
\arguments{
\item{.data}{a dataframe which may be grouped}
}
\value{
the .data dataframe with the history graph removed
}
\description{
This is unlikely to be needed directly and is mostly and internal function
}
\examples{
library(dplyr)
library(dtrackr)
mtcars \%>\% track() \%>\% comment("A comment") \%>\% p_clear() \%>\% history()
}
