% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_count_if}
\alias{p_count_if}
\title{Simple count_if dplyr summary function}
\usage{
p_count_if(..., na.rm = TRUE)
}
\arguments{
\item{...}{expression to be evaluated}

\item{na.rm}{ignore NA values?}
}
\value{
a count of the number of times the expression evaluated to true, in the current context
}
\description{
Simple count_if dplyr summary function
}
\examples{
library(dplyr)
library(dtrackr)
tmp = iris \%>\% dplyr::group_by(Species)
tmp \%>\% dplyr::summarise(long_ones = p_count_if(Petal.Length > 4))
}
