% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_pause}
\alias{p_pause}
\title{Pause tracking the dataframe}
\usage{
p_pause(.data)
}
\arguments{
\item{.data}{\itemize{
\item a tracked dataframe
}}
}
\value{
the .data dataframe with history graph tracking paused
}
\description{
Pause tracking the dataframe
}
\examples{
library(dplyr)
iris \%>\% track() \%>\% pause() \%>\% history()
}
