% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_set}
\alias{p_set}
\title{Set the dtrackr history graph}
\usage{
p_set(.data, .graph)
}
\arguments{
\item{.data}{a dataframe which may be grouped}

\item{.graph}{a history graph list (consisting of nodes, edges, and head) see examples}
}
\value{
the .data dataframe with the history graph metadata set to the provided value
}
\description{
This is unlikely to be useful to an end user and is called automatically by many of the other
functions here. On the off chance you need to copy history metadata from one dataframe to another
}
\examples{
library(dplyr)
library(dtrackr)
mtcars \%>\% p_set(iris \%>\% comment("A comment") \%>\% p_get()) \%>\% history()
}
