% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{p_summarise}
\alias{p_summarise}
\title{Summarise a data set}
\usage{
p_summarise(.data, ..., .messages = "", .headline = "", .tag = NULL)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{
  Arguments passed on to \code{\link[dplyr:summarise]{dplyr::summarise}}
  \describe{
    \item{\code{.by}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, a selection of columns to
group by for just this operation, functioning as an alternative to \code{\link[dplyr:group_by]{group_by()}}. For
details and examples, see \link[dplyr:dplyr_by]{?dplyr_by}.}
    \item{\code{.groups}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Grouping structure of the
result.
\itemize{
\item "drop_last": dropping the last level of grouping. This was the
only supported option before version 1.0.0.
\item "drop": All levels of grouping are dropped.
\item "keep": Same grouping structure as \code{.data}.
\item "rowwise": Each row is its own group.
}

When \code{.groups} is not specified, it is chosen
based on the number of rows of the results:
\itemize{
\item If all the results have 1 row, you get "drop_last".
\item If the number of rows varies, you get "keep" (note that returning a
variable number of rows was deprecated in favor of \code{\link[dplyr:reframe]{reframe()}}, which
also unconditionally drops all levels of grouping).
}

In addition, a message informs you of that choice, unless the result is ungrouped,
the option "dplyr.summarise.inform" is set to \code{FALSE},
or when \code{summarise()} is called from a function in a package.}
  }}

\item{.messages}{a set of glue specs. The glue code can use any summary
variable defined in the ... parameter, or any global variable, or
\{.strata\}}

\item{.headline}{a headline glue spec. The glue code can use any summary
variable defined in the ... parameter, or any global variable, or
\{.strata\}}

\item{.tag}{if you want the summary data from this step in the future then
give it a name with .tag.}
}
\value{
the .data dataframe summarised with the history graph updated showing
the summarise operation as a new stage
}
\description{
Summarising a data set acts in the normal \code{dplyr} manner to collapse groups
to individual rows. Any columns resulting from the summary can be added to
the history graph In the history this also joins any stratified branches and
allows you to generate some summary statistics about the un-grouped data. See
\code{\link[dplyr:summarise]{dplyr::summarise()}}.
}
\examples{
library(dplyr)
library(dtrackr)

tmp = iris \%>\% group_by(Species)
tmp \%>\% summarise(avg = mean(Petal.Length), .messages="{avg} length") \%>\% history()
}
\seealso{
dplyr::summarise()
}
