% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{plot.trackr_graph}
\alias{plot.trackr_graph}
\title{Plots a history graph as html}
\usage{
\method{plot}{trackr_graph}(x, fill = "lightgrey", fontsize = "8", colour = "black", ...)
}
\arguments{
\item{x}{a dtrackr history graph (e.g. output from \code{\link[=history]{history()}})}

\item{fill}{the default node fill colour}

\item{fontsize}{the default font size}

\item{colour}{the default font colour}

\item{...}{not used}
}
\value{
HTML displayed
}
\description{
Plots a history graph as html
}
\examples{
library(dplyr)
library(dtrackr)
iris \%>\% comment("hello {.total} rows") \%>\% history() \%>\% plot()
}
