% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ungroup}
\alias{summarise}
\alias{mutate}
\alias{transmute}
\alias{select}
\alias{relocate}
\alias{rename}
\alias{rename_with}
\alias{arrange}
\alias{pivot_wider}
\alias{pivot_longer}
\alias{group_by}
\alias{distinct}
\alias{filter}
\alias{group_modify}
\alias{inner_join}
\alias{left_join}
\alias{right_join}
\alias{full_join}
\alias{semi_join}
\alias{anti_join}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_modify}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{relocate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{rename_with}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{ungroup}}}

  \item{tidyr}{\code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{pivot_wider}}}
}}

