% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrackr.R
\name{resume}
\alias{resume}
\title{Resume tracking the dataframe. This may reset the grouping of the tracked data}
\usage{
resume(.data)
}
\arguments{
\item{.data}{\itemize{
\item a tracked dataframe
}}
}
\value{
the .data dataframe with history graph tracking resumed
}
\description{
Resume tracking the dataframe. This may reset the grouping of the tracked data
}
\examples{
library(dplyr)
iris \%>\% track() \%>\% pause() \%>\% resume() \%>\% history()
}
