% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{save_dot}
\alias{save_dot}
\title{Save DOT to a file}
\usage{
save_dot(
  dot,
  filename,
  size = std_size$half,
  maxWidth = size$width,
  maxHeight = size$height,
  rot = size$rot,
  formats = c("dot", "png", "pdf", "svg")
)
}
\arguments{
\item{dot}{- a graphviz dot string}

\item{filename}{- the root of the desired filename (minus extension)}

\item{size}{- a list of length and width in inches e.g. a std_size}

\item{maxWidth}{- a width in inches is size is not defined}

\item{maxHeight}{- a heught in inches if size is not defined}

\item{rot}{- an angle of rotation for the saved file if size is not defined}

\item{formats}{- some of "pdf","dot","svg","png","ps"}
}
\value{
the SVG as a string
}
\description{
Convert a digraph and save as a file
}
\examples{
dot2svg("digraph {A->B}")
}
