% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{save_dot}
\alias{save_dot}
\title{Save DOT content to a file}
\usage{
save_dot(
  dot,
  filename,
  size = std_size$half,
  maxWidth = size$width,
  maxHeight = size$height,
  formats = c("dot", "png", "pdf", "svg"),
  ...
)
}
\arguments{
\item{dot}{a graphviz dot string}

\item{filename}{the full path of the filename (minus extension for multiple formats)}

\item{size}{a list of length and width in inches e.g. a std_size}

\item{maxWidth}{a width in inches is size is not defined}

\item{maxHeight}{a height in inches if size is not defined}

\item{formats}{some of "pdf","dot","svg","png","ps"}

\item{...}{ignored}
}
\value{
a list with items \code{paths} with the absolute paths of the saved files as a named list, and \code{svg} as the SVG string of the rendered dot file.
}
\description{
Convert a digraph in dot file to SVG and save it to an output file
}
\examples{
save_dot("digraph {A->B}",tempfile())
}
